Moved Gift type to shared/types.ts. Fixed imports
