+++ components/DeleteModal.tsx
@@ -1,6 +1,6 @@
 import { TitleText } from './TitleText';
 import React, { Dispatch, SetStateAction } from 'react';
-import { Gift } from '~/pages';
+import { Gift } from '~/shared/types';
 import { Modal } from './Modal';
 import { Button } from './Button';
 import { SvgCheckMarkIcon } from '~/icons/CheckMarkIcon';
