+++ shared/types.ts
@@ -1,0 +1,7 @@
+export type Gift = {
+  receiver: string;
+  gift: string;
+  uuid: string;
+  createdAt?: Date;
+  updatedAt?: Date;
+};
