+++ utils/giftRequests.ts
@@ -1,5 +1,6 @@
 import axios from 'axios';
-import { CreateGift, Gift } from '~/pages';
+import { CreateGift } from '~/pages';
+import { Gift } from '~/shared/types';
 
 const baseURL = '/api/gifts';
